changed return int to double on divide method

+++ src/main/java/com/example/springdocker/service/MyMathCalc.java
@@ -17,7 +17,7 @@ public class MyMathCalc {
         return a*b;
     }
 
-    public float divide(int a, int b){
+    public double divide(int a, int b){
         if(b!=0){
             return a/b;
         } else {
